+++ src/pages/logout.tsx
@@ -1,7 +1,10 @@
 import { Hono } from "hono";
 import { deleteCookie } from "hono/cookie";
+import { csrf } from "hono/csrf";
 
 const logout = new Hono();
+
+logout.use(csrf());
 
 logout.post("/", async (c) => {
   await deleteCookie(c, "login");